Allow exporting of baresip contacts to Android contacts

+++ app/src/main/kotlin/com/tutpro/baresip/ContactListAdapter.kt
@@ -1,5 +1,6 @@
 package com.tutpro.baresip
 
+import android.Manifest
 import android.app.Activity
 import android.content.Context
 import android.content.DialogInterface
@@ -126,9 +127,10 @@ class ContactListAdapter(private val ctx: Context, private val rows: ArrayList<C
                                 Log.e("Baresip", "Could not delete file '$id.img")
                             }
                         }
-                        Contact.contacts().removeAt(position)
-                        if (contact.androidContact)
+                        if (contact.androidContact &&
+                                Utils.checkPermission(ctx, Manifest.permission.WRITE_CONTACTS))
                             ContactActivity.deleteAndroidContact(ctx, contact)
+                        Contact.contacts().removeAt(position)
                         Contact.save()
                         this.notifyDataSetChanged()
                     }
